Adcionado o método criarCartas()
