+++ project/assets/material/carta-frente-padrao.tres
@@ -1,40 +1,37 @@
-[gd_resource type="ShaderMaterial" load_steps=11 format=2]
+[gd_resource type="ShaderMaterial" load_steps=10 format=2]
 
-[ext_resource path="res://assets/texture/carta-padrao.png" type="Texture" id=1]
+[ext_resource path="res://assets/texture/numeros/padrao/um_n.png" type="Texture" id=1]
+[ext_resource path="res://assets/texture/numeros/padrao/um_d_c.png" type="Texture" id=2]
 
-[sub_resource type="VisualShaderNodeTexture" id=1]
-texture = ExtResource( 1 )
+[sub_resource type="VisualShaderNodeTextureUniform" id=1]
+uniform_name = "carta_d_c"
+color_default = 1
 
-[sub_resource type="VisualShaderNodeScalarFunc" id=2]
-function = 31
+[sub_resource type="VisualShaderNodeTextureUniform" id=2]
+uniform_name = "carta_n"
+texture_type = 2
 
 [sub_resource type="VisualShaderNodeColorUniform" id=3]
 uniform_name = "carta_cor"
 
 [sub_resource type="VisualShaderNodeVectorOp" id=4]
-operator = 6
+output_port_for_preview = 0
 
 [sub_resource type="VisualShaderNodeScalarUniform" id=5]
 uniform_name = "carta_metalico"
 
-[sub_resource type="VisualShaderNodeScalarOp" id=6]
-output_port_for_preview = 0
-operator = 2
-
-[sub_resource type="VisualShaderNodeScalarUniform" id=7]
+[sub_resource type="VisualShaderNodeScalarUniform" id=6]
 uniform_name = "carta_rugosidade"
 
-[sub_resource type="VisualShaderNodeScalarOp" id=8]
-operator = 2
-
-[sub_resource type="VisualShader" id=9]
+[sub_resource type="VisualShader" id=7]
 code = "shader_type spatial;
 render_mode specular_schlick_ggx;
 
+uniform sampler2D carta_d_c : hint_black;
 uniform vec4 carta_cor : hint_color;
-uniform sampler2D tex_frg_2;
 uniform float carta_metalico;
 uniform float carta_rugosidade;
+uniform sampler2D carta_n : hint_normal;
 
 
 
@@ -44,37 +41,42 @@ void vertex() {
 }
 
 void fragment() {
+// TextureUniform:10
+	vec3 n_out10p0;
+	float n_out10p1;
+	{
+		vec4 n_tex_read = texture(carta_d_c, UV.xy);
+		n_out10p0 = n_tex_read.rgb;
+		n_out10p1 = n_tex_read.a;
+	}
+
 // ColorUniform:4
 	vec3 n_out4p0 = carta_cor.rgb;
 	float n_out4p1 = carta_cor.a;
 
-// Texture:2
-	vec4 tex_frg_2_read = texture(tex_frg_2, UV.xy);
-	vec3 n_out2p0 = tex_frg_2_read.rgb;
-	float n_out2p1 = tex_frg_2_read.a;
-
-// ScalarFunc:3
-	float n_out3p0 = 1.0 - dot(n_out2p0, vec3(0.333333, 0.333333, 0.333333));
-
 // VectorOp:5
-	vec3 n_out5p0 = max(n_out4p0, vec3(n_out3p0));
+	vec3 n_out5p0 = n_out10p0 + n_out4p0;
 
 // ScalarUniform:6
 	float n_out6p0 = carta_metalico;
 
-// ScalarOp:7
-	float n_out7p0 = n_out3p0 * n_out6p0;
-
 // ScalarUniform:8
 	float n_out8p0 = carta_rugosidade;
 
-// ScalarOp:9
-	float n_out9p0 = n_out3p0 * n_out8p0;
+// TextureUniform:11
+	vec3 n_out11p0;
+	float n_out11p1;
+	{
+		vec4 n_tex_read = texture(carta_n, UV.xy);
+		n_out11p0 = n_tex_read.rgb;
+		n_out11p1 = n_tex_read.a;
+	}
 
 // Output:0
 	ALBEDO = n_out5p0;
-	METALLIC = n_out7p0;
-	ROUGHNESS = n_out9p0;
+	METALLIC = n_out6p0;
+	ROUGHNESS = n_out8p0;
+	NORMAL = n_out11p0;
 
 }
 
@@ -83,28 +85,26 @@ void light() {
 
 }
 "
-graph_offset = Vector2( -381.924, -57 )
-nodes/fragment/0/position = Vector2( 400, 320 )
-nodes/fragment/2/node = SubResource( 1 )
-nodes/fragment/2/position = Vector2( -320, 160 )
-nodes/fragment/3/node = SubResource( 2 )
-nodes/fragment/3/position = Vector2( -80, 260 )
+graph_offset = Vector2( 22.5, 92 )
+nodes/fragment/0/position = Vector2( 1140, 140 )
 nodes/fragment/4/node = SubResource( 3 )
-nodes/fragment/4/position = Vector2( -320, 40 )
+nodes/fragment/4/position = Vector2( 240, 220 )
 nodes/fragment/5/node = SubResource( 4 )
-nodes/fragment/5/position = Vector2( 140, 180 )
+nodes/fragment/5/position = Vector2( 500, 120 )
 nodes/fragment/6/node = SubResource( 5 )
-nodes/fragment/6/position = Vector2( -80, 360 )
-nodes/fragment/7/node = SubResource( 6 )
-nodes/fragment/7/position = Vector2( 140, 320 )
-nodes/fragment/8/node = SubResource( 7 )
-nodes/fragment/8/position = Vector2( -100, 440 )
-nodes/fragment/9/node = SubResource( 8 )
-nodes/fragment/9/position = Vector2( 120, 560 )
-nodes/fragment/connections = PoolIntArray( 2, 0, 3, 0, 3, 0, 5, 1, 4, 0, 5, 0, 5, 0, 0, 0, 3, 0, 7, 0, 6, 0, 7, 1, 7, 0, 0, 2, 3, 0, 9, 0, 8, 0, 9, 1, 9, 0, 0, 3 )
+nodes/fragment/6/position = Vector2( 820, 200 )
+nodes/fragment/8/node = SubResource( 6 )
+nodes/fragment/8/position = Vector2( 820, 280 )
+nodes/fragment/10/node = SubResource( 1 )
+nodes/fragment/10/position = Vector2( 220, 20 )
+nodes/fragment/11/node = SubResource( 2 )
+nodes/fragment/11/position = Vector2( 820, 400 )
+nodes/fragment/connections = PoolIntArray( 11, 0, 0, 7, 10, 0, 5, 0, 4, 0, 5, 1, 5, 0, 0, 0, 8, 0, 0, 3, 6, 0, 0, 2 )
 
 [resource]
-shader = SubResource( 9 )
+shader = SubResource( 7 )
 shader_param/carta_cor = Color( 1, 1, 0, 1 )
-shader_param/carta_metalico = 0.0
-shader_param/carta_rugosidade = 1.0
+shader_param/carta_metalico = 0.4
+shader_param/carta_rugosidade = 0.7
+shader_param/carta_d_c = ExtResource( 2 )
+shader_param/carta_n = ExtResource( 1 )
