+++ project/assets/texture/numeros/padrao/coringa_d_c.png
