+++ project/scene/Carta.tscn
@@ -1,9 +1,11 @@
-[gd_scene load_steps=3 format=2]
+[gd_scene load_steps=4 format=2]
 
 [ext_resource path="res://assets/material/carta-frente-padrao.tres" type="Material" id=1]
 [ext_resource path="res://assets/meshes/carta.obj" type="ArrayMesh" id=2]
+[ext_resource path="res://src/views/CartaView.cs" type="Script" id=3]
 
 [node name="Carta" type="Spatial"]
+script = ExtResource( 3 )
 
 [node name="mesh" type="MeshInstance" parent="."]
 mesh = ExtResource( 2 )
